split jwplayer
